+++ java/src/test/java/io/cucumber/query/NamingStrategyAcceptanceTest.java
@@ -23,7 +23,7 @@ import java.util.List;
 import java.util.Map;
 
 import static io.cucumber.query.Jackson.OBJECT_MAPPER;
-import static io.cucumber.query.Repository.RepositoryFeature.INCLUDE_GHERKIN_DOCUMENT;
+import static io.cucumber.query.Repository.RepositoryFeature.INCLUDE_GHERKIN_DOCUMENTS;
 import static io.cucumber.query.NamingStrategy.ExampleName.NUMBER_AND_PICKLE_IF_PARAMETERIZED;
 import static io.cucumber.query.NamingStrategy.ExampleName.PICKLE;
 import static io.cucumber.query.NamingStrategy.FeatureName.EXCLUDE;
@@ -86,7 +86,7 @@ public class NamingStrategyAcceptanceTest {
 
     private static Repository createRepository() {
         return Repository.builder()
-                .feature(INCLUDE_GHERKIN_DOCUMENT, true)
+                .feature(INCLUDE_GHERKIN_DOCUMENTS, true)
                 .build();
     }